study(step3): Dealer, Gamer 파일 분리 및 makeReport abstract 함수로 분리
